+++ public/stylesheets/style.sass
@@ -1,8 +1,8 @@
-@import ../bootstrap/scss/bootstrap.scss
+// @import ../bootstrap/scss/bootstrap.scss
 
-@import url(https://use.fontawesome.com/releases/v5.8.1/css/all.css)
-@import url(//netdna.bootstrapcdn.com/font-awesome/3.2.1/css/font-awesome.css)
-@import url('https://fonts.googleapis.com/css?family=Dancing+Script')
+// @import url(https://use.fontawesome.com/releases/v5.8.1/css/all.css)
+// @import url(//netdna.bootstrapcdn.com/font-awesome/3.2.1/css/font-awesome.css)
+// @import url('https://fonts.googleapis.com/css?family=Dancing+Script')
 
 body
   font: 14px "Lucida Grande", Helvetica, Arial, sans-serif
