--- routes/coach.js
@@ -1,9 +1,0 @@
-var express = require('express');
-var router = express.Router();
-
-module.exports = router;
-
-// $( ".userimg-coach" ).click(function goToURL() {
-//     location.href='/trainingTypes');
-//   });
-
