--- routes/coachees.js
@@ -1,4 +1,0 @@
-var express = require('express');
-var router = express.Router();
-
-module.exports = router;
