--- routes/middleware/auth.js
@@ -1,11 +1,0 @@
-
-module.exports = {
-    isAuthenticated: (req, res, next) => {
-        if (req.session.usuario) {
-          next();
-        }
-        else{
-            res.redirect('/login');
-        }
-    }
-};
