+++ src/global.sass
@@ -1,0 +1,5 @@
+@import ../bootstrap/scss/bootstrap.scss
+
+@import url(https://use.fontawesome.com/releases/v5.8.1/css/all.css)
+@import url(//netdna.bootstrapcdn.com/font-awesome/3.2.1/css/font-awesome.css)
+@import url('https://fonts.googleapis.com/css?family=Dancing+Script')
