+++ src/pages/alunos/login/styles.sass
@@ -1,0 +1,8 @@
+
+.Livel
+  color: white
+  font-size: 150px
+  text-align: center
+  margin: 130px
+  font-family: Dancing Script
+  font-weight: bold
