+++ src/pages/alunos/trainingType/trainingTypes.hbs
