+++ trash/index.hbs
@@ -1,0 +1,12 @@
+{{!-- <body class="home">
+    <div class="txt-over-Img">
+      <img src="/images/FundoIndex.png" class="image-index"></img>
+      <div class="centered Livel-index">Livel
+        <div class="btnlogin">
+          <button id="singupbutton" onclick="location.href='/login'"
+            class="btnchangelogin btn btn-primary btn-block">LOGIN</button>
+        </div>
+      </div>
+    </div>
+  
+  </body> --}}
