+++ trash/routes/auth.js
@@ -1,0 +1,11 @@
+
+// module.exports = {
+//     isAuthenticated: (req, res, next) => {
+//         if (req.session.usuario) {
+//           next();
+//         }
+//         else{
+//             res.redirect('/login');
+//         }
+//     }
+// };
