+++ trash/routes/coach.js
@@ -1,0 +1,10 @@
+// var express = require('express');
+// var router = express.Router();
+
+// module.exports = router;
+
+// // $( ".userimg-coach" ).click(function goToURL() {
+// //     location.href='/trainingTypes');
+// //   });
+
+
