+++ trash/routes/users.js
@@ -1,0 +1,9 @@
+// var express = require('express');
+// var router = express.Router();
+
+// /* GET users listing. */
+// router.get('/', function(req, res, next) {
+//   res.send('respond with a resource');
+// });
+
+// module.exports = router;
